+++ src/components/MyRoutines.jsx
@@ -1,6 +1,5 @@
-import React, { useState, useEffect } from "react";
-import { createRoutines } from "../api/myRoutines";
-import useAuth from "../hooks/useAuth";
+import React from "react";
+import CreatingRoutines from "./CreatingRoutines";
 /**
  *   As a registered user on the My Routines tab, I want to:
  *     [x] - be shown a form to create a new routine
@@ -14,51 +13,13 @@ import useAuth from "../hooks/useAuth";
  *                   on the routine
  *              [] - be able to remove any activity from the routine
  */
-const CreatingRoutines = () => {
-  const { token, user } = useAuth();
-  const [title, setTitle] = useState("");
-  const [goal, setGoal] = useState("");
-  const [routines, setRoutines] = useState("");
 
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const routineObj = { name: title, goal };
-    console.log(routineObj, "Making sure that the routineObj is connected");
-    const response = await createRoutines(routineObj, token);
-    console.log(response, "Getting response?");
-    const newRoutine = response.data;
-  };
-
-  return (
-    <div className="newRoutines">
-      <form onSubmit={handleSubmit}>
-        <input
-          type="text"
-          value={title}
-          onChange={(e) => setTitle(e.target.value)}
-          placeholder="title"
-        />
-        <input
-          type="text"
-          value={goal}
-          onChange={(e) => setGoal(e.target.value)}
-          placeholder="Goal"
-        />
-        <button type="submit">Create Routines</button>
-      </form>
-    </div>
-  );
-};
 // get Routines again and filter by comparing user.username to creator name
 // to only show the user's routines
 const getFilteredRoutines = () => {
   return;
 };
 const MyRoutines = ({ post, setPost }) => {
-  // const [title, setTitle] = useState("");
-  // const [description, setDescription] = useState("");
-  //const [routines, setRoutines] = useState("");
-
   return (
     <div>
       <h1> My Routines</h1>
